app/requirements.txt
autodynatrace==1.0.81
autowrapt==1.0
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
oneagent-sdk==1.4.0.20210127.165413
python-dotenv==0.18.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
